rename the notifications package to notification
Rename the package no code changes beyond that. Since moving items and
transforms under the notifications package there is no need for it to be
plural.
