+++ src/main/java/com/blackducksoftware/integration/hub/dataservices/notification/items/NotificationItemCount.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.dataservices.notifications.items;
+package com.blackducksoftware.integration.hub.dataservices.notification.items;
 
 import com.blackducksoftware.integration.hub.api.notification.NotificationItem;
