+++ src/main/java/com/blackducksoftware/integration/hub/dataservices/notification/items/VulnerabilityContentItem.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.dataservices.notifications.items;
+package com.blackducksoftware.integration.hub.dataservices.notification.items;
 
 import java.util.List;
 import java.util.UUID;
