+++ src/main/java/com/blackducksoftware/integration/hub/dataservices/notification/transforms/AbstractPolicyTransform.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.dataservices.notifications.transforms;
+package com.blackducksoftware.integration.hub.dataservices.notification.transforms;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -19,8 +19,8 @@ import com.blackducksoftware.integration.hub.api.component.ComponentVersionStatu
 import com.blackducksoftware.integration.hub.api.notification.NotificationItem;
 import com.blackducksoftware.integration.hub.api.policy.PolicyRule;
 import com.blackducksoftware.integration.hub.api.project.ProjectVersion;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationContentItem;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.PolicyNotificationFilter;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.NotificationContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.PolicyNotificationFilter;
 import com.blackducksoftware.integration.hub.exception.BDRestException;
 import com.blackducksoftware.integration.hub.exception.HubItemTransformException;
 import com.blackducksoftware.integration.hub.exception.MissingUUIDException;
