+++ src/main/java/com/blackducksoftware/integration/hub/dataservices/notification/transforms/VulnerabilityTransform.java
@@ -1,4 +1,4 @@
-package com.blackducksoftware.integration.hub.dataservices.notifications.transforms;
+package com.blackducksoftware.integration.hub.dataservices.notification.transforms;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,8 +15,8 @@ import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotif
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationItem;
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilitySourceQualifiedId;
 import com.blackducksoftware.integration.hub.api.project.ProjectVersion;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationContentItem;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.VulnerabilityContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.NotificationContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.VulnerabilityContentItem;
 import com.blackducksoftware.integration.hub.exception.HubItemTransformException;
 import com.blackducksoftware.integration.hub.util.HubUrlParser;
