+++ src/test/java/com/blackducksoftware/integration/hub/dataservices/notifications/items/NotificationCountDataTest.java
@@ -7,7 +7,7 @@ import java.util.Date;
 
 import org.junit.Test;
 
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.NotificationCountData;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.NotificationCountData;
 
 public class NotificationCountDataTest {
