+++ src/test/java/com/blackducksoftware/integration/hub/dataservices/notifications/items/PolicyOverrideItemTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 
 import com.blackducksoftware.integration.hub.api.policy.PolicyRule;
 import com.blackducksoftware.integration.hub.api.project.ProjectVersion;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.PolicyOverrideContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.PolicyOverrideContentItem;
 
 public class PolicyOverrideItemTest {
