+++ src/test/java/com/blackducksoftware/integration/hub/dataservices/notifications/items/VulnerabilityItemCountTest.java
@@ -8,7 +8,7 @@ import org.mockito.Mockito;
 
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationContent;
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilityNotificationItem;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.VulnerabilityItemCount;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.VulnerabilityItemCount;
 
 public class VulnerabilityItemCountTest {
