+++ src/test/java/com/blackducksoftware/integration/hub/dataservices/notifications/items/VulnerabilityItemTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 
 import com.blackducksoftware.integration.hub.api.notification.VulnerabilitySourceQualifiedId;
 import com.blackducksoftware.integration.hub.api.project.ProjectVersion;
-import com.blackducksoftware.integration.hub.dataservices.notifications.items.VulnerabilityContentItem;
+import com.blackducksoftware.integration.hub.dataservices.notification.items.VulnerabilityContentItem;
 
 public class VulnerabilityItemTest {
